Add UTF-8 rendering support for TTY, curses, and dumplogs.
Decided to add legit UTF-8 support to EvilHack. Modeled this mainly off
of the UnNetHack implementation, as NetHack 3.7's method is so
entrenched with it's newer symbols code, it wouldn't really translate
over. Did cherry-pick a bit of code from 3.7 (proper ncursew API for
handling wide characters). Made improvements compared to the UnNetHack
implementation (specifically simpler UTF-8 detection, and transparent
IBMgraphics-to-Unicode conversation) while also creating
EvilHack-specific bits, such as full dumplog support (both text and HTML
dumplogs, these now render fully using unicode characters), and if you
prefer the IBMgraphics or DECgraphics look/feel, you can use UTF-8
translation (default for pretty much every terminal) instead of having
to switch to CP437 - it will render as you'd expect.

Probably a bit more to do here, but this is a good initial commit for
it. In my opinion this is a marked improvement over 3.7's enhanced/UTF-8
mode, as it's version dumbs down to using ascii-style corridors and open
doors, sinks are still rendered as ascii but fountains are properly
rendered. It also uses a triple tilde to render pools of water and lava
- while that looks good, not a lot of fonts have a triple tilde, so
you're limited to what you can use without having a bunch of boxes with
questions marks inside of them. I did borrow the unicode symbols 3.7
uses for altars, graves, and trees (also applies to S_TREE class
monsters), and made magic chests use the house symbol which actually
works pretty well for that.

To use: make sure your terminal is using UTF-8 translation, and put
'OPTIONS=symset:UTF8graphics' in your config while commenting out any
other symsets. Works in both TTY and curses mode.

The details of all the file edits for the technical-minded:

* src/unicode.c: New file. CP437-to-Unicode and DEC-to-Unicode lookup
  tables, pututf8char() encoder, utf8str_from_codepoint() buffer
  encoder, detect_utf8_terminal() via nl_langinfo(CODESET).
* include/global.h: Widen nhsym from uchar to int to hold Unicode
  codepoints (showsyms grows from ~199 to ~796 bytes).
* include/flag.h: Add iflags.UTF8graphics and iflags.supports_utf8.
* include/rm.h: Add H_UTF8 symbol handling constant.
* include/extern.h: Declare unicode.c and get_utf8_sym() functions.
* src/drawing.c: Add "UTF8" to known_handling[], add utf8_graphics[]
  default array (MAXPCHARS Unicode codepoints for terrain symbols),
  init_utf8_graphics() loader called from switch_symbols(), and
  get_utf8_sym() accessor for dumplog use.
* src/options.c: Add UTF8graphics boolean option with mutual exclusion
  against IBMgraphics/DECgraphics. Add U+NNNN hex codepoint parsing
  to sym_val(). Fix sign extension bug: (int)*buf -> (int)(uchar)*buf
  so CP437 bytes 0x80-0xFF are not mangled by signed char promotion.
* dat/symbols: Add UTF8graphics symset with box-drawing, shading, and
  monster class overrides (S_ent=Psi, S_plant=pound, S_orb=degree).
* win/tty/wintty.c: g_putch() checks iflags.supports_utf8 and routes
  through pututf8char(), converting IBMgraphics/DECgraphics bytes to
  Unicode codepoints transparently.
* win/curses/curswins.c: write_char() uses ncursesw mvwadd_wch() with
  setcchar()/wchar_t for codepoints above 0x7F. Add NCURSES_WIDECHAR
  define and wchar.h include.
* win/curses/cursmain.c: Add setlocale(LC_ALL, "") before initscr().
* win/curses/cursinit.c: Remove stale locale.h include.
* sys/unix/unixmain.c: Call setlocale(LC_ALL, "") before initoptions()
  and detect_utf8_terminal() after initoptions().
* sys/unix/Makefile.src: Add unicode.c/unicode.o to build. Link
  -lncursesw instead of -lncurses for wide character support.
* sys/unix/hints/linux-debug: Link -lncursesw instead of -lncurses.
* src/detect.c: dump_map() uses pointer-based buffer (COLNO*4+1) for
  multi-byte UTF-8 output. Uses get_utf8_sym() for terrain and direct
  UTF-8 encoding for non-ASCII monster class symbols.
* src/end.c: Save/restore showsyms[] around init_symbols() so monster
  class symbol overrides (ents, orbs, mounds) survive into dumplogs.
* src/windows.c: Extend htmlsym[] with Unicode codepoints for HTML
  dumplogs. Add ch > 0x7F fallback for monster class symbols.
* src/pager.c: Farlook prefix uses utf8str_from_codepoint() for
  Unicode symbols instead of truncating via encglyph().
* src/mapglyph.c: decode_mixed() encodes multi-byte codepoints via
  utf8str_from_codepoint() instead of single-byte truncation.
* src/files.c: Handle iflags.UTF8graphics in config file processing.

+++ include/global.h
@@ -87,9 +87,10 @@ typedef xchar boolean; /* 0 or 1 */
 #endif
 
 /*
- * type nhsym: loadable symbols go into this type
+ * type nhsym: loadable symbols go into this type,
+ * widened from uchar to int for UTF-8/Unicode codepoint support
  */
-typedef uchar nhsym;
+typedef int nhsym;
 
 #ifndef STRNCMPI
 #ifndef __SASC_60 /* SAS/C already shifts to stricmp */
@@ -390,6 +391,5 @@ struct savefile_info {
 #ifndef nethack_enter
 #define nethack_enter(argc, argv) ((void) 0)
 #endif
-
 
 #endif /* GLOBAL_H */
